requirements.txt
appnope==0.1.0
click==6.6
decorator==4.0.10
Django==1.10.6
django-annoying==0.10.3
django-cleanup==1.0.0
django-password-reset==1.0
django-resized==0.3.5
django-yarn==0.0.1
Flask==0.11.1
image==1.5.7
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
munch==2.1.0
nexmo==1.2.0
olefile==0.44
pexpect==4.2.0
pickleshare==0.7.3
Pillow==4.2.1
prompt-toolkit==1.0.3
ptyprocess==0.5.1
Pygments==2.1.3
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.11.10
